Remove unused codes
