+++ common/storages/fuse/src/io/write/mod.rs
@@ -12,15 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod block_stream_writer;
+mod block_compactor;
 mod block_writer;
 mod meta_writer;
 mod segment_writer;
 
 // for testing only
-pub use block_stream_writer::BlockCompactor;
-pub use block_stream_writer::BlockStreamWriter;
-pub use block_stream_writer::SegmentInfoStream;
+pub use block_compactor::BlockCompactor;
 pub use block_writer::write_block;
 pub use block_writer::write_data;
 pub use block_writer::BlockWriter;
