+++ common/storages/index/src/range_filter.rs
@@ -37,9 +37,7 @@ use common_planners::RequireColumnsVisitor;
 pub struct ClusterKeyInfo {
     pub cluster_key_id: u32,
     pub cluster_key_index: Vec<usize>,
-    pub exprs: Vec<Expression>,
     pub expression_executor: Option<ExpressionExecutor>,
-    pub data_schema: DataSchemaRef,
 }
 
 #[derive(Clone)]
